requirements.txt
aiogram~=3.0.0b7
asyncio~=3.4.3
moviepy~=1.0.3
yt_dlp~=2023.7.6
requests~=2.27.1
python-dotenv~=0.20.0